권장휴가 관련 파일 삭제 - holiday로 통합 및 holidayType을 추가하여 관리하도록 변경 DTO 생성 및 수정 - 기존 Controller에서 사용하던 dto 삭제 및 DTO 폴더 생성 후 각 dto를 사용하여 변환하도록 수정

+++ src/main/java/com/lshdainty/myhr/service/HolidayService.java
@@ -1,15 +1,15 @@
 package com.lshdainty.myhr.service;
 
 import com.lshdainty.myhr.domain.Holiday;
-import com.lshdainty.myhr.domain.Recommend;
+import com.lshdainty.myhr.domain.HolidayType;
 import com.lshdainty.myhr.repository.HolidayRepository;
-import com.lshdainty.myhr.repository.RecommendRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -17,37 +17,48 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class HolidayService {
     private final HolidayRepository holidayRepository;
-    private final RecommendRepository recommendRepository;
 
     @Transactional
-    public Long save(String name, String date) {
-        Holiday holiday = Holiday.createHoliday(name, date);
+    public Long save(String name, String date, HolidayType type) {
+        Holiday holiday = Holiday.createHoliday(name, date, type);
         holidayRepository.save(holiday);
         return holiday.getSeq();
     }
 
     public Holiday findHoliday(Long seq) {
-        return holidayRepository.findHoliday(seq);
+        Holiday holiday = holidayRepository.findHoliday(seq);
+
+        // holiday 없으면 에러 반환
+        if (Objects.isNull(holiday)) { throw new IllegalArgumentException("holiday not found"); }
+
+        return holiday;
     }
 
     public List<Holiday> findHolidays() {
         return holidayRepository.findHolidays();
     }
 
+    public List<Holiday> findHolidaysByType(HolidayType type) {
+        return holidayRepository.findHolidaysByType(type);
+    }
+
     @Transactional
-    public void editHoliday(Long holidaySeq, String name, String date) {
+    public void editHoliday(Long holidaySeq, String name, String date, HolidayType type) {
         Holiday findHoliday = holidayRepository.findHoliday(holidaySeq);
-        findHoliday.updateHoliday(name, date);
+
+        // holiday 없으면 에러 반환
+        if (Objects.isNull(findHoliday)) { throw new IllegalArgumentException("holiday not found"); }
+
+        findHoliday.updateHoliday(name, date, type);
     }
 
     @Transactional
     public void deleteHoliday(Long holidaySeq) {
         Holiday findHoliday = holidayRepository.findHoliday(holidaySeq);
-        holidayRepository.delete(findHoliday);
-    }
 
-    // 권장휴무는 단순 조회만 함
-    public List<Recommend> findRecommends() {
-        return recommendRepository.findRecommends();
+        // holiday 없으면 에러 반환
+        if (Objects.isNull(findHoliday)) { throw new IllegalArgumentException("holiday not found"); }
+
+        holidayRepository.delete(findHoliday);
     }
 }
